Write a program that prints the first 50 Fibonacci numbers, starting with 1 and 2, followed by a new line.

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -11,7 +11,7 @@
 
 int main(void)
 {
-	int i, a, b, c;
+	long long i, a, b, c;
 	
 	a = 1;
 	b = 1;
